Terminology update. Change 'frame' to 'block' where appropriate.
Alsaplayer used to use the work 'frame' to describe a block of N (N > 1)
stereo sample pairs. The MPEG encoder/decoder and and speech related DSP
algorithms also use 'frame' this way.

However, Jack, Alsa, libsndfile, Apple's CoreAudio and others use the word
'frame' to describe the M samples (where M is the number of channels) that
belong to the same sampling instance.

With this change we leave change Alsaplayer to agree with Jack, Alsa etc.